separated 'add cluster form' into a different component
